zalo_tts/requirements.txt
PyAudio==0.2.11
python-dotenv==0.17.0